Make ToCTCP handle any type of CTCP message, move DCC handling to separate file
