+++ pkg/irc/conn.go
@@ -223,6 +223,11 @@ func (c *Client) recv() {
 				c.setNick(msg.LastParam())
 			}
 
+		case Privmsg:
+			if ctcp := msg.ToCTCP(); ctcp != nil {
+				c.handleCTCP(ctcp)
+			}
+
 		case ReplyWelcome:
 			c.setNick(msg.Params[0])
 			c.setRegistered(true)
@@ -242,10 +247,17 @@ func (c *Client) recv() {
 
 		}
 
-		if ctcp := msg.ToCTCP(); ctcp != nil {
-			go c.Download(ctcp)
-		}
-
 		c.Messages <- msg
+	}
+}
+
+func (c *Client) handleCTCP(ctcp *CTCP) {
+	switch ctcp.Command {
+	case "DCC":
+		if strings.HasPrefix(ctcp.Params, "SEND") {
+			if dccSend := ParseDCCSend(ctcp); dccSend != nil {
+				go c.Download(dccSend)
+			}
+		}
 	}
 }
